requirements.txt
av>=10.0.0
numpy>=1.24.0
Pillow>=10.0.0
pysdl2>=0.9.16
pysdl2-dll>=2.28.0